* Fixed the problem.  Added some additional logic to enable full folder conversion. * Still not done, but it works, you can view zones now, and it won't crash when converting all RoF2 zones. * Known issues are that plants, trees, campfires, torches, and stuff like that won't convert, but those appear to be frag types 23 and 24 that we don't support yet.

+++ FileConverter/Entities/Zone.cs
@@ -60,24 +60,6 @@ namespace OpenEQ.FileConverter.Entities
             });
         }
 
-        //    def coalesceObjectMeshes(self):
-        //    for obj in self.objects:
-        //        startmeshcount = len(obj.meshes)
-        //        matmeshes = {}
-        //        for mesh in obj.meshes:
-        //            mat = mesh.material.filenames, mesh.material.flags
-        //            if mat not in matmeshes:
-        //                matmeshes[mat] = []
-        //    matmeshes[mat].append(mesh)
-        //        obj.meshes = []
-        //    poss = 0
-        //        for meshlist in matmeshes.values():
-        //            if len(meshlist) == 1:
-        //                obj.meshes.append(meshlist[0])
-        //                continue
-        //            gmesh = reduce(lambda a, b: a.add(b), meshlist)
-        //            obj.meshes += gmesh.optimize()
-
         public void CoalesceObjectMeshes()
         {
             foreach (var obj in ZoneObjects)
@@ -219,7 +201,7 @@ namespace OpenEQ.FileConverter.Entities
                                 bw.Write(mesh.Polygons.Count);
                                 foreach (var poly in mesh.Polygons)
                                 {
-                                    bw.WriteFloatArray(poly.Item2.to_array());
+                                    bw.WriteIntArray(poly.Item2.to_array());
                                 }
                                 foreach (var poly in mesh.Polygons)
                                 {
